+++ Mosaic_editor/Classes/Puzzle.cs
@@ -22,16 +22,22 @@ namespace Mosaic_editor.Classes
         public string filename { get; private set; }
 
         public int GridSpacing = 100;
+        private ColourPalette palette;
 
         public int cols { get; private set; }
         public int rows { get; private set; }
 
-        public string Name = "A First Puzzle";
-        public string Difficulty = "Hard";
+        public string name = "A First Puzzle";
+        public string difficulty = "Hard";
+        public int repeats = 3;
+        public bool dontMangle = false;
+        public bool fixedColors = false;
 
         // Limit the maximum number of hexagons
         const int WIDTH_LIMIT = 50;
         const int HEIGHT_LIMIT = 50;
+
+        public List<Color> colors = new List<Color>();
 
         internal Hexagon selectedHexagon = null;
 
@@ -42,11 +48,12 @@ namespace Mosaic_editor.Classes
         /// <param name="width">in pixels</param>
         /// <param name="height">in pixels</param>
         /// <param name="gridSpacing">in pixels</param>
-        public Puzzle(int width, int height, int gridSpacing)
+        public Puzzle(int width, int height, int gridSpacing, ColourPalette palette)
         {
             this.width = width;
             this.height = height;
             this.GridSpacing = gridSpacing;
+            this.palette = palette;
 
             CreateHexArray(width, height);
             refreshPositions();
@@ -55,67 +62,34 @@ namespace Mosaic_editor.Classes
         /// <summary>
         /// Creates a list of hexagons enough to fill the given area
         /// </summary>
-        /// <param name="width"></param>
-        /// <param name="height"></param>
-        private void CreateHexArray(int width, int height)
+        /// <param name="windowWidth"></param>
+        /// <param name="windowHeight"></param>
+        private void CreateHexArray(int windowWidth, int windowHeight)
         {
             HexagonList = new List<Hexagon>();
 
-            // Work out how many rows and columns there are
-            this.cols = (int)Math.Floor(width / (this.GridSpacing * 1.5));
+            // Work out how many rows and columns fit into this window
+            const int MARGINS = 20;
+            var unitsAcross = (windowWidth - MARGINS) / this.GridSpacing;
+            this.cols = (int)Math.Floor((unitsAcross - 0.5) / 1.5);
+
             var triangleHeight = this.GridSpacing * Constants.COS30;
-            this.rows = (int)Math.Floor((height - triangleHeight) / triangleHeight);
+            var unitsDown = (windowHeight - MARGINS) / triangleHeight;
+            this.rows = (int)Math.Floor(unitsDown - 1);
             Console.WriteLine($"this puzzle has {cols} columns and {rows} rows");
 
             // row and col are simple counters
             int col0 = 1;
             for (int row = 0; row < rows; row++)
             {
-                /// x = x0 + GridSpacing;
                 for (int col = col0; col < cols; col += 2)
                 {
-                    var hex = new Hexagon(row, col); // , x, y, GridSpacing);
-                    // HexagonArray[x, y] = hex;
+                    var hex = new Hexagon(row, col);
+                    hex.parent = this;
                     HexagonList.Add(hex);
-                    //x += dx;
-                    //if (x + GridSpacing > width) break;
                 }
-                //x0 = (int)(GridSpacing * 1.5) - x0;   // this handles the alternating row offsets
-                //y += dy;
-                //if (y + GridSpacing > height) break;
                 col0 = 1 - col0;
             }
-
-            // x and y are the positions on screen in pixels of the (centres of) the hexagons
-            // dx is the spacing between hex centres across the screen
-            // dy is the spacing between hex centres down the screen
-            // EVEN rows (0, 2, 4, ..) start from x = gridSpacing
-            // ODD rows (1, 3, 5, ..) start from x = 1.5 times the gridSpacing
-
-            //int dx = 3 * GridSpacing;
-            //int dy = (int)(GridSpacing * Constants.COS30);
-            //int x;
-            //int x0 = (int)(GridSpacing * 1.5); // the starting offset for each row
-
-            //int y = dy;
-            //int col0 = 1;
-            //for (int row = 0; row < HEIGHT_LIMIT; row++)
-            //{
-            //    x = x0 + GridSpacing;
-
-            //    for (int col = col0; col < WIDTH_LIMIT; col += 2)
-            //    {
-            //        var hex = new Hexagon(row, col, x, y, GridSpacing);
-            //        // HexagonArray[x, y] = hex;
-            //        HexagonList.Add(hex);
-            //        x += dx;
-            //        if (x + GridSpacing > width) break;
-            //    }
-            //    x0 = (int)(GridSpacing * 1.5) - x0;   // this handles the alternating row offsets
-            //    y += dy;
-            //    if (y + GridSpacing > height) break;
-            //    col0 = 1 - col0;
-            //}
         }
 
         /// <summary>
@@ -125,6 +99,11 @@ namespace Mosaic_editor.Classes
         internal void refresh()
         {
             HexagonList.ForEach(h => h.refresh());
+        }
+
+        internal Color getColor(int colorNo)
+        {
+            return palette.getColor(colorNo);
         }
 
         /// <summary>
@@ -154,6 +133,7 @@ namespace Mosaic_editor.Classes
         /// <param name="g"></param>
         internal void draw(Graphics g)
         {
+            g.TextRenderingHint = System.Drawing.Text.TextRenderingHint.AntiAlias;
             foreach (var hex in HexagonList)
             {
                 hex.draw(g);
@@ -171,10 +151,13 @@ namespace Mosaic_editor.Classes
         /// <summary>
         /// TODO
         /// </summary>
-        internal static Puzzle load(ColourPalette palette)
+        internal static Puzzle load() // ColourPalette palette)
         {
             var dlg = new OpenFileDialog();
-            dlg.InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            var myDocuments = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            dlg.InitialDirectory = Path.Combine(myDocuments, "mosaic");
+            if (!Directory.Exists(dlg.InitialDirectory)) Directory.CreateDirectory(dlg.InitialDirectory);
+
             dlg.Filter = "JSON files|*.json";
             if (dlg.ShowDialog() == DialogResult.OK)
             {
@@ -185,8 +168,9 @@ namespace Mosaic_editor.Classes
                     {
                         var result = JsonConvert.DeserializeObject(text, typeof(Puzzle)) as Puzzle;
                         result.RepairLinks();
-                        Console.WriteLine($"Loaded puzzle from {dlg.FileName}");
                         result.refreshPositions();
+                        result.recreatePalette();
+                        Console.WriteLine($"Loaded puzzle from {dlg.FileName}");
                         return result;
                     }
                     catch (Exception ex)
@@ -202,27 +186,139 @@ namespace Mosaic_editor.Classes
             // a change
 
             // Failed!  Return a new empty puzzle
-            var newPuzzle = new Puzzle(10, 10, 100);
-            newPuzzle.Name = "Puzzle failed to load";
+            var newPuzzle = new Puzzle(10, 10, 100, new ColourPalette(6));
+            newPuzzle.name = "Puzzle failed to load";
             return newPuzzle;
         }
 
+        /// <summary>
+        /// After a load, recreate the current palette
+        /// </summary>
+        private void recreatePalette()
+        {
+            this.palette = new ColourPalette(colors);
+        }
+
+        internal Range2D getActiveRange()
+        {
+            var result = new Range2D();
+
+            var activeHexagons = HexagonList.Where(h => h.isActive);    // get a list of the active ones
+            if (activeHexagons.Count() > 0)
+            {
+                var colMin = activeHexagons.Select(h => h.col).Min();
+                var colMax = activeHexagons.Select(h => h.col).Max();
+                var rowMin = activeHexagons.Select(h => h.row).Min();
+                var rowMax = activeHexagons.Select(h => h.row).Max();
+
+                result = new Range2D(colMin, rowMin, colMax, rowMax);
+            }
+
+            return result;
+        }
+        
+        internal Range2D getWindowRange()
+        {
+            if (HexagonList.Count > 0)
+            {
+                var colMin = HexagonList.Select(h => h.col).Min();
+                var colMax = HexagonList.Select(h => h.col).Max();
+                var rowMin = HexagonList.Select(h => h.row).Min();
+                var rowMax = HexagonList.Select(h => h.row).Max();
+                return new Range2D(colMin, rowMin, colMax, rowMax);
+            }
+            else
+            {
+                return new Range2D(); // empty rectangle
+            }
+        }
+
+        internal void translate(int x, int y)
+        {
+            var activeHexagons = HexagonList.Where(h => h.isActive);    // get a list of the active ones
+            foreach(var h in activeHexagons)
+            {
+                h.col += x;
+                h.row += y;
+            }
+            refreshPositions();
+        }
+
+        /// <summary>
+        /// Recentres the puzzle in the existing window
+        /// </summary>
+        /// <param name="width">Window width in pixels</param>
+        /// <param name="height">Window height in pixels</param>
+        /// <param name="gridSize">Triangle width in pixels</param>
         internal void recentre(int width, int height, int gridSize = 0)
         {
             this.width = width;
             this.height = height;
-            if (gridSize > 0) this.GridSpacing = gridSize;
+            if (gridSize > 0) this.GridSpacing = gridSize;  // ..otherwise it remains unchanged from its previous value
+
             var activeHexagons = HexagonList.Where(h => h.isActive);    // save a list of the active ones
+            var bounds = getActiveRange();   // the bounds of the active puzzle
+            Console.WriteLine($"current puzzle bounds: {bounds.x1},{bounds.y1} - {bounds.x2},{bounds.y2}");
+
+            // Reinitialise the HexagonList with new hexagons
             CreateHexArray(width, height);
-            // re-apply the active ones into the new list
-            foreach(var hex in activeHexagons)
+            var window = getWindowRange(); // the size of the window
+            window.Dump("current window bounds");
+            /// Console.WriteLine($"(current window bounds: {window.Left},{window.Top} - {window.Right},{window.Bottom})");
+
+            // Offset the puzzle to put it in the centre of the screen.
+            int xOffset = (int)((window.width - bounds.width) / 2);     // round to nearest int
+            int yOffset = (int)((window.height - bounds.height) / 2);
+            Console.WriteLine($"..recentre() puts the window centre at {xOffset},{yOffset}");
+            // Adjust to the nearest valid hex position
+            if ((yOffset & 1) == 1)
             {
-                Console.WriteLine($"Active: {hex.col}, {hex.row}");
+                // Its in an odd row, so must be in an even column
+                if ((xOffset & 1) == 1)
+                {
+                    xOffset = (xOffset > 0) ? xOffset - 1 : xOffset + 1;
+                }
+            }
+            else
+            {
+                // Its in an even row, so has to be in an odd column
+                if ((xOffset & 1) == 0)
+                {
+                    xOffset = (xOffset > 0) ? xOffset - 1 : xOffset + 1;
+                }
+            }
+            Console.WriteLine($"..recentre() puts the corrected window centre at {xOffset},{yOffset}");
+
+            xOffset -= bounds.x1;
+            yOffset -= bounds.y1;
+            // NOTE: Not all offsets are valid!  Row 0 only has columns 1, 3, 5, ... and row 1 has columns 0, 2, 4, ... etc.
+            // Valid offsets include:
+            //      1,-1 1,+1 3,-1 3+1 etc.
+            //      2,0 4,0 6,0 etc..
+            // i.e. if xOffset is odd, yOffset must also be odd.
+            //      if xOffset is even, yOffset must also be even.
+
+            if ((xOffset & 1) != (yOffset & 1)) {
+                // One is odd and one is even.  Fix!
+                yOffset = (yOffset > 0) ? yOffset - 1 : yOffset + 1;
+            }
+            Console.WriteLine($"recentre will offset this puzzle by {xOffset},{yOffset}");
+
+            // re-apply the active ones into the new list
+            foreach (var hex in activeHexagons)
+            {
+                hex.col += xOffset;
+                hex.row += yOffset;
+                Console.WriteLine($"Hex moved to: {hex.col}, {hex.row}");
                 var match = HexagonList.FirstOrDefault(hex2 => hex2.col == hex.col && hex2.row == hex.row);
                 if (match != null)
                 {
                     HexagonList.Add(hex);
                     HexagonList.Remove(match);
+                }
+                else
+                {
+                    Console.WriteLine($"..found no place for tile[{hex.col},{hex.row}]!");
                 }
             }
             refreshPositions();
@@ -240,6 +336,8 @@ namespace Mosaic_editor.Classes
         {
             foreach(var hex in HexagonList)
             {
+                hex.parent = this;
+
                 foreach(var t in hex.triangles)
                 {
                     t.parent = hex;
@@ -250,39 +348,70 @@ namespace Mosaic_editor.Classes
         /// <summary>
         /// TODO
         /// </summary>
-        internal void save(ColourPalette palette)
+        internal void save()
         {
             var activeHexagons = HexagonList.Where(h => h.isActive);    // LINQ
 
             if (activeHexagons.Count() == 0) return;
 
-            var colMin = activeHexagons.Select(h => h.col).Min();
-            var colMax = activeHexagons.Select(h => h.col).Max();
-            var rowMin = activeHexagons.Select(h => h.row).Min();
-            var rowMax = activeHexagons.Select(h => h.row).Max();
+            var bounds = this.getActiveRange();
 
-            Console.WriteLine($"crop: {colMin},{rowMin} - {colMax},{ rowMax}");
+            //Console.WriteLine($"crop: {colMin},{rowMin} - {colMax},{ rowMax}");
+            bounds.Dump();
+            int xOffset = bounds.x1;
+            int yOffset = bounds.y1;
+            if ((yOffset & 1) != (yOffset & 1))
+            {
+                // MUST both be odd or both even!  Not the case this time, so fix it:
+                yOffset = yOffset < 1 ? yOffset++ : yOffset--;
+            }
+            if (xOffset == 0 && yOffset == 0)   // [0,0] is invalid
+            {
+                xOffset++;
+                // yOffset++;
+            }
+            Console.WriteLine($"save() is moving the puzzle [-{xOffset},-{yOffset}]");
+            // this.translate(-xOffset, -yOffset);  // align the active hexagons to top left corner
 
-            var usedHexagons = HexagonList.Where(h => h.row >= rowMin && h.row <= rowMax && h.col >= colMin && h.col <= colMax);
+            bounds = this.getActiveRange();
+            bounds.Dump("after aligning bounds");
+            // Console.WriteLine($"after aligning bounds: {bounds.Left},{bounds.Top} - {bounds.Right},{bounds.Bottom}");
+
+            // var usedHexagons = HexagonList.Where(h => h.row >= rowMin && h.row <= rowMax && h.col >= colMin && h.col <= colMax);
+            var usedHexagons = HexagonList.Where(h => bounds.contains(h)); // h.row >= bounds.Top && h.row < bounds.Bottom && h.col >= bounds.Left && h.col < bounds.Right);
 
             var puzzle =(Puzzle)this.MemberwiseClone();
+
+            puzzle.colors = this.palette.getColors();
 
             // puzzle.Name = this.Name;
             // puzzle.Difficulty = this.Difficulty;
             puzzle.DateCreated= DateTime.Now;
             puzzle.DateUpdated = DateTime.Now;
-            puzzle.HexagonList = usedHexagons.ToList();
-            var json = JsonConvert.SerializeObject(puzzle, Formatting.Indented);
+            puzzle.HexagonList = activeHexagons.ToList(); // usedHexagons.ToList();
 
             var dlg = new SaveFileDialog();
             dlg.Filter = "JSON files|*.json";
-            dlg.InitialDirectory = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            var myDocuments = Environment.GetFolderPath(Environment.SpecialFolder.MyDocuments);
+            dlg.InitialDirectory = Path.Combine(myDocuments, "mosaic");
+            if (!Directory.Exists(dlg.InitialDirectory)) Directory.CreateDirectory(dlg.InitialDirectory);
             dlg.FileName = "mosaic puzzles.json";
             if (dlg.ShowDialog() == DialogResult.OK)
             {
                 puzzle.filename = dlg.FileName;
+
+                //var settings = new JsonSerializerSettings { 
+                //    Formatting = Formatting.Indented,
+                //    MaxDepth = 20,
+                //    ReferenceLoopHandling = ReferenceLoopHandling.Ignore
+                //};
+                var json = JsonConvert.SerializeObject(puzzle, Formatting.Indented);
+
                 Console.WriteLine($"Saving to {dlg.FileName}");
                 File.WriteAllText(dlg.FileName, json);
+
+                var merger = new PuzzleMerger();
+                merger.run(dlg.FileName, "puzzleSet1.js");   // tell the merger where the last puzzle was saved
             }
 
         }